use build tag `hyperscan_v52` for literal support
