+++ hyperscan/literal_v52.go
@@ -1,5 +1,5 @@
-//go:build !hyperscan_v4
-// +build !hyperscan_v4
+//go:build hyperscan_v52 || hyperscan_v54
+// +build hyperscan_v52 hyperscan_v54
 
 package hyperscan
